requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fastapi==0.95.1
filetype==1.2.0
flatbuffers==23.3.3
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.28.1
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.13.1
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
lazy_loader==0.2
llvmlite==0.39.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
numba==0.56.4
numpy==1.23.5
onnxruntime==1.14.1
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
pooch==1.7.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==11.0.0
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
PyMatting==1.1.8
Pympler==1.0.1
pyreadline3==3.4.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==306
pyzmq==25.0.2
rembg==2.0.35
requests==2.29.0
rich==13.3.5
scikit-image==0.20.0
scipy==1.10.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
streamlit==1.22.0
sympy==1.11.1
tenacity==8.2.2
tifffile==2023.4.12
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.22.0
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
yarl==1.9.2
zipp==3.15.0
